+++ server/src/main/java/com/bside/bside_311/dto/AttachDto.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @Getter
 public class AttachDto {
   @Schema(example = "1", description = "첨부 번호")
-  private Integer attachNo;
+  private Long attachNo;
   @Schema(example = "http://www.naver.com", description = "프로필 사진 링크.")
   private String attachUrl;
   @Schema(example = "profile", description = "첨부 타입")
